+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.6] - 2026-01-27
+
+### Changed
+
+- Updated all optional properties in property type definitions to accept `null` values in addition to `undefined`
+  - Affected interfaces: `PropertyBase`, `PropertyString`, `PropertyNumber`, `PropertyBoolean`, `PropertyObject`, `PropertyArray`, `PropertyCredentialId`, `PropertyEncryptedString`, `FilterOperators`, `RootFilter`
+
 ## [0.1.5] - 2026-01-27
 
 ### Changed
@@ -99,6 +106,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added project structure documentation
 - Added development guidelines and contributing guide
 
+[0.1.6]: https://github.com/choice-open/atomemo-plugin-schema/compare/v0.1.5...v0.1.6
 [0.1.5]: https://github.com/choice-open/atomemo-plugin-schema/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/choice-open/atomemo-plugin-schema/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/choice-open/atomemo-plugin-schema/compare/v0.1.2...v0.1.3